Community/server/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
pillow==11.3.0
psycopg==3.2.9
PyJWT==2.8.0
python-dotenv==1.1.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
typing_extensions==4.14.1
gunicorn==23.0.0